fruit/requirements.txt
streamlit==1.26.0
numpy==1.24.4
tensorflow==2.10.0  # Python 3.9 호환
Pillow==9.4.0
gdown==4.7.1
setuptools